requirements.txt
langgraph
langchain_openai 
langchain_experimental 
pymupdf 
tiktoken 
qdrant-client
chainlit==1.1.302
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.9
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.1
python-dotenv==1.0.1
faiss-cpu